backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
user-agents
email-validator
httpx